refactor(lolla): extrair VenuePoiList e venueMapData de VenueMap
