+++ components/landings/lollapalooza/VenueMap.tsx
@@ -1,8 +1,9 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { renderToStaticMarkup } from 'react-dom/server';
 import L from 'leaflet';
-import { MapPin, Navigation, Plane, Building2, Trees, CarFront, Info, ChevronRight, Utensils, TrainFront } from 'lucide-react';
-import { optimizeRemoteImageUrl } from '../../../data/mediaConfig';
+import { MapPin } from 'lucide-react';
+import { POIS, getPoiStyle } from './venueMapData';
+import { VenuePoiList } from './VenuePoiList';
 
 const VenueMap: React.FC = () => {
   const mapContainerRef = useRef<HTMLDivElement>(null);
@@ -14,74 +15,6 @@ const VenueMap: React.FC = () => {
   const itemsRef = useRef<(HTMLButtonElement | null)[]>([]);
 
   const [activeIndex, setActiveIndex] = useState<number | null>(null);
-
-  // Dados dos Pontos de Interesse
-  const pois = [
-    {
-      name: "Aeroporto de Guarulhos (GRU)",
-      coords: [-23.426173, -46.467733] as [number, number],
-      type: 'airport',
-      distance: '55 km',
-      time: '1h 30min',
-      description: "Principal porta de entrada internacional. Recomendamos nosso transfer exclusivo que busca grupos acima de 10 pessoas mediante agendamento.",
-      transitInfo: "Expresso Aeroporto até Luz ➔ Linha 4-Amarela até Pinheiros ➔ Linha 9-Esmeralda até Estação Autódromo. (~2h)",
-      image: "images/lollapalooza/venue/gru.jpg"
-    },
-    {
-      name: "Aeroporto de Congonhas (CGH)",
-      coords: [-23.627325, -46.656545] as [number, number],
-      type: 'airport',
-      distance: '15 km',
-      time: '40 min',
-      description: "Aeroporto central para voos domésticos. Localização estratégica, a apenas 15 minutos dos hotéis parceiros na região da Berrini.",
-      transitInfo: "Uber até Estação Campo Belo (Linha 5-Lilás) ➔ Santo Amaro ➔ Linha 9-Esmeralda até Estação Autódromo. (~55min)",
-      image: "images/lollapalooza/venue/cgh.jpg"
-    },
-    {
-      name: "Parque Ibirapuera",
-      coords: [-23.587416, -46.657634] as [number, number],
-      type: 'landmark',
-      distance: '18 km',
-      time: '45 min',
-      description: "O pulmão verde de SP. Ótimo para um passeio relaxante na manhã pré-festival. A região conta com hotéis de alto padrão.",
-      transitInfo: "Estação Moema (Linha 5-Lilás) ➔ Santo Amaro ➔ Linha 9-Esmeralda até Estação Autódromo. (~50min)",
-      image: "images/lollapalooza/venue/ibirapuera.jpg"
-    },
-    {
-      name: "Av. Paulista",
-      coords: [-23.561349, -46.656388] as [number, number],
-      type: 'landmark',
-      distance: '22 km',
-      time: '50 min',
-      description: "O coração financeiro e cultural. Oferecemos pacotes com hospedagem nesta região para quem quer curtir a cidade além dos shows.",
-      transitInfo: "Estação Consolação (Linha 2-Verde) ➔ Pinheiros (Linha 4) ➔ Linha 9-Esmeralda até Estação Autódromo. (~1h)",
-      image: "images/lollapalooza/venue/avenida-paulista.jpg"
-    },
-    {
-      name: "Pinheiros (Bares e Restaurantes)",
-      coords: [-23.566374, -46.702966] as [number, number],
-      type: 'landmark',
-      distance: '20 km',
-      time: '45 min',
-      description: "Bairro boêmio com a melhor gastronomia e vida noturna. Perfeito para o 'esquenta' ou para jantar após o festival.",
-      transitInfo: "Estação Pinheiros (Linha 9-Esmeralda) ➔ Direto até Estação Autódromo. Rota mais rápida de trem. (~35min)",
-      image: "images/lollapalooza/venue/pinheiros.jpg"
-    }
-  ];
-
-  // Helper para determinar estilo do ícone baseado no tipo/nome
-  const getPoiStyle = (poi: typeof pois[0]) => {
-    if (poi.type === 'airport') {
-      return { Icon: Plane, bgColor: '#4B5563', iconColor: 'white', label: 'Aeroporto' };
-    }
-    if (poi.name.includes('Parque')) {
-      return { Icon: Trees, bgColor: '#059669', iconColor: 'white', label: 'Natureza' };
-    }
-    if (poi.name.includes('Pinheiros')) {
-      return { Icon: Utensils, bgColor: '#FFD600', iconColor: '#003B8E', label: 'Gastronomia' };
-    }
-    return { Icon: Building2, bgColor: '#0056D2', iconColor: 'white', label: 'Cidade' };
-  };
 
   useEffect(() => {
     let a11yTimer: ReturnType<typeof setTimeout>;
@@ -155,7 +88,7 @@ const VenueMap: React.FC = () => {
       // --- ÍCONES DOS POIS ---
       const markers: L.Marker[] = [interlagosMarker];
 
-      pois.forEach((poi, index) => {
+      POIS.forEach((poi, index) => {
         const style = getPoiStyle(poi);
         const iconSvgString = renderToStaticMarkup(
           <style.Icon size={18} color={style.iconColor} strokeWidth={2.5} />
@@ -219,7 +152,7 @@ const VenueMap: React.FC = () => {
 
     if (newIndex !== null && mapInstanceRef.current && markersRef.current[index]) {
       const marker = markersRef.current[index];
-      const poi = pois[index];
+      const poi = POIS[index];
 
       mapInstanceRef.current.flyTo(poi.coords, 14, { animate: true, duration: 1.5 });
       marker.openPopup();
@@ -242,136 +175,14 @@ const VenueMap: React.FC = () => {
 
         <div className="bg-white rounded-3xl shadow-2xl overflow-hidden flex flex-col md:flex-row">
 
-          {/* Informações / Lista de Distâncias */}
-          {/* Aumentei o max-h no mobile de 50vh para 600px para evitar cortes no card */}
-          <div className="w-full md:w-1/3 p-6 md:p-8 bg-white flex flex-col border-r border-gray-100 h-auto flex-shrink-0">
-            <div className="inline-flex items-center gap-2 text-anhanga-blue font-bold uppercase tracking-wider text-xs mb-3">
-              <MapPin size={16} aria-hidden="true" /> Localização & Referências
-            </div>
-
-            <h3 className="text-2xl font-black text-anhanga-darkBlue mb-2 leading-tight">
-              Explore os arredores de <br /><span className="text-anhanga-yellow">Interlagos</span>
-            </h3>
-
-            <p className="text-gray-500 text-sm mb-6">
-              Clique nos pontos abaixo para ver detalhes.
-            </p>
-
-            {/* Container da Lista com Scroll - Adicionado padding extra (p-2) para evitar corte da sombra/scale */}
-            <div
-              ref={listContainerRef}
-              className="flex-1 overflow-y-auto p-4 custom-scrollbar flex flex-col gap-4 min-h-0 -mx-4"
-            >
-              {pois.map((poi, idx) => {
-                const style = getPoiStyle(poi);
-                return (
-                  <button
-                    key={poi.name}
-                    ref={el => {
-                      itemsRef.current[idx] = el;
-                    }}
-                    onClick={() => handleListClick(idx)}
-                    className={`w-full text-left transition-all duration-300 rounded-xl border-2 group overflow-hidden relative ${activeIndex === idx
-                      ? 'border-anhanga-blue shadow-xl ring-2 ring-blue-100 z-10'
-                      : 'border-gray-100 hover:border-anhanga-yellow hover:shadow-md'
-                      }`}
-                  >
-                    {/* Imagem de Fundo Otimizada */}
-                    <div className="absolute inset-0 z-0">
-                      <img
-                        src={optimizeRemoteImageUrl(poi.image, 400, 300)}
-                        alt={`Local: ${poi.name}`}
-                        width="400"
-                        height="300"
-                        loading="lazy"
-                        decoding="async"
-                        className="w-full h-full object-cover transition-transform duration-700 group-hover:scale-110 opacity-30"
-                      />
-                      {/* Overlay para garantir legibilidade do texto sobre a imagem */}
-                      <div className={`absolute inset-0 transition-colors duration-300 ${activeIndex === idx ? 'bg-blue-50/90' : 'bg-white/90 group-hover:bg-white/80'
-                        }`}></div>
-                    </div>
-
-                    {/* Conteúdo com z-10 para ficar sobre a imagem */}
-                    <div className="relative z-10 p-3 flex items-center justify-between">
-                      <div className="flex items-center gap-3">
-                        <div
-                          className="size-8 rounded-full flex items-center justify-center shadow-sm transition-colors shrink-0"
-                          style={{
-                            backgroundColor: activeIndex === idx ? '#0056D2' : style.bgColor,
-                            color: activeIndex === idx ? 'white' : style.iconColor
-                          }}
-                        >
-                          <style.Icon size={14} />
-                        </div>
-                        <div className="min-w-0">
-                          <p className={`font-bold text-sm leading-tight transition-colors ${activeIndex === idx ? 'text-anhanga-darkBlue whitespace-normal' : 'text-gray-800 truncate'}`}>
-                            {poi.name}
-                          </p>
-                          <p className="text-[10px] text-gray-500 uppercase tracking-wider font-semibold">
-                            {style.label}
-                          </p>
-                        </div>
-                      </div>
-                      <div className="text-right flex flex-col items-end shrink-0 pl-2">
-                        {activeIndex === idx ? (
-                          <ChevronRight size={16} className="text-anhanga-blue rotate-90 transition-transform" />
-                        ) : (
-                          <>
-                            <p className="font-black text-anhanga-darkBlue text-sm">{poi.distance}</p>
-                            <p className="text-xs text-gray-500 flex items-center justify-end gap-1">
-                              <CarFront size={10} /> {poi.time}
-                            </p>
-                          </>
-                        )}
-                      </div>
-                    </div>
-
-                    {/* Detalhes Expandidos */}
-                    <div className={`relative z-10 grid transition-all duration-300 ease-in-out ${activeIndex === idx ? 'grid-rows-[1fr] opacity-100' : 'grid-rows-[0fr] opacity-0'
-                      }`}>
-                      <div className="overflow-hidden min-h-0">
-                        <div className="p-3 pt-0 text-sm text-gray-600 border-t border-blue-100/50 mx-3 mt-1 space-y-3">
-
-                          {/* Descrição Geral */}
-                          <div className="flex items-start gap-2 pt-3">
-                            <Info size={14} className="text-anhanga-blue shrink-0 mt-0.5" />
-                            <p className="leading-snug">{poi.description}</p>
-                          </div>
-
-                          {/* Rota de Transporte Público */}
-                          <div className="bg-white/60 backdrop-blur-sm p-2.5 rounded-lg border border-gray-200">
-                            <div className="flex items-center gap-2 mb-1.5 text-gray-800 font-bold text-xs uppercase tracking-wide">
-                              <TrainFront size={12} className="text-gray-600" />
-                              Rota até o Autódromo (Transporte Público):
-                            </div>
-                            <p className="text-xs text-gray-600 leading-snug pl-5 border-l-2 border-gray-300">
-                              {poi.transitInfo}
-                            </p>
-                          </div>
-
-                          {/* Rodapé do Item */}
-                          <div className="flex items-center justify-between text-xs font-semibold text-anhanga-darkBlue bg-blue-50/50 p-2 rounded-lg backdrop-blur-sm">
-                            <span className="flex items-center gap-1"><CarFront size={12} /> Distância: {poi.distance}</span>
-                            <span>Tempo est. carro: {poi.time}</span>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  </button>
-                );
-              })}
-            </div>
-
-            <div className="mt-4 pt-4 border-t border-gray-100 hidden md:block">
-              <div className="flex items-start gap-3 bg-blue-50 p-3 rounded-lg">
-                <Navigation className="text-anhanga-blue shrink-0 mt-0.5" size={16} />
-                <p className="text-xs text-blue-800 leading-snug">
-                  <span className="font-bold">Dica:</span> Selecione um ponto para ver no mapa.
-                </p>
-              </div>
-            </div>
-          </div>
+          <VenuePoiList
+            pois={POIS}
+            activeIndex={activeIndex}
+            listContainerRef={listContainerRef}
+            itemsRef={itemsRef}
+            onItemClick={handleListClick}
+            getPoiStyle={getPoiStyle}
+          />
 
           {/* Mapa */}
           <div className="w-full md:w-2/3 h-[400px] md:h-auto min-h-[400px] relative bg-gray-200">
